Partial fix for server <-> client save time ahead of world time bug
Anything that saves the current world time and is saved on a player ship can have a time in the future saved when traveling between servers or after a universe reset. In vanilla evolving animals, animals that produce items, and plants can suffer from this. Plants appear to use hardcoded functions and cannot be fixed this way.